update request URLs and app settings
