+++ lib/javascripts/base_app.js
@@ -177,14 +177,6 @@ BaseApp.prototype = {
     return url;
   },
 
-  queryCustomer: function() {
-    var self = this;
-    this.switchTo('requesting');
-    this.zafClient.get('ticket.requester.email').then(function(data) {
-      self.ajax('getProfile', data["ticket.requester.email"]);
-    });
-  },
-
   isZatEnabled: function() {
     return (this.id() === 0);
   }
